Refactor webpack configs into partials
