+++ config/webpack/partials/babel-loader.js
@@ -1,0 +1,18 @@
+"use strict";
+
+var partial = require("webpack-partial").partial;
+
+module.exports = function (config) {
+  return partial(config, {
+    module: {
+      loaders: [
+        {
+          name: "babel",
+          test: /\.jsx?$/,
+          exclude: [/node_modules/],
+          loaders: [require.resolve("babel-loader") + "?optional=runtime"]
+        }
+      ]
+    }
+  });
+};
