+++ config/webpack/partials/binary-loader.js
@@ -1,0 +1,17 @@
+"use strict";
+
+var partial = require("webpack-partial").partial;
+
+module.exports = function (config) {
+  return partial(config, {
+    module: {
+      loaders: [
+        {
+          name: "binary",
+          test: /\.(png|svg|woff|woff2|ttf|eot|jpg|jpeg)$/i,
+          loader: require.resolve("url-loader") + "?limit=10000"
+        }
+      ]
+    }
+  });
+};
