+++ config/webpack/partials/source-map.js
@@ -1,0 +1,9 @@
+"use strict";
+
+var partial = require("webpack-partial").partial;
+
+module.exports = function (config) {
+  return partial(config, {
+    devtool: "source-map"
+  });
+};
