+++ config/webpack/webpack.config.test.js
@@ -1,35 +1,18 @@
 "use strict";
+
 /**
  * Webpack frontend test configuration.
  */
-var path = require("path");
-var _ = require("lodash");
-var ExtractTextPlugin = require("extract-text-webpack-plugin");
-var prodCfg = require("./webpack.config");
 
-// Replace with `__dirname` if using in project root.
-var ROOT = process.cwd();
+var partial = require("webpack-partial").partial;
+var base = require("./partials/base");
+var baseLoader = require("./partials/base-loader");
+var testEntry = require("./partials/test-entry");
+var testOutput = require("./partials/test-output");
 
-module.exports = {
-  cache: true,
-  context: path.join(ROOT, "test/client"),
-  entry: "./main",
-  output: {
-    path: __dirname,
-    filename: "main.js",
-    publicPath: "/assets/"
-  },
-  resolve: _.merge({}, prodCfg.resolve, {
-    alias: {
-      // Allow root import of `client/FOO` from ROOT/client.
-      client: path.join(ROOT, "client")
-    }
-  }),
-  module: prodCfg.module,
-  devtool: "source-map",
-  plugins: [
-    new ExtractTextPlugin("style.css", {
-      allChunks: true
-    })
-  ]
-};
+module.exports = partial(
+  base,
+  baseLoader,
+  testEntry,
+  testOutput
+);
